fix output arg
