+++ cmd/tools/cmq/gc.go
@@ -53,7 +53,7 @@ func gc(c *cli.Context) error {
 		return nil
 	}
 
-	mc, err := newMetadataClient(getOpts(cliContext))
+	mc, err := newMetadataClient()
 
 	if err != nil {
 		fmt.Errorf("newMetadataClient error: %v", err)
