+++ cmd/tools/cmq/meta.go
@@ -2,8 +2,12 @@ package main
 
 import (
 	"fmt"
-	"github.com/gocql/gocql"
+	"io/ioutil"
 	"time"
+
+	"github.com/gocql/gocql"
+	"github.com/urfave/cli"
+	"gopkg.in/yaml.v2"
 )
 
 type opts struct {
@@ -30,7 +34,9 @@ const (
 	numConns        = 1
 )
 
-func newMetadataClient(opts *opts) (*metadataClient, error) {
+func newMetadataClient() (*metadataClient, error) {
+
+	opts := getOpts(cliContext)
 
 	cluster := gocql.NewCluster(opts.Hosts) // TODO: add support for multiple hosts
 
@@ -148,4 +154,90 @@ func (m *metadataClient) Query(table string, f filter) (row map[string]interface
 	*/
 
 	return
+}
+
+// ZoneConfig has config for cassandra for each zone
+type ZoneConfig struct {
+	Hosts    string `yaml:"hosts"` // TODO: add support for multiple hosts
+	Port     int    `yaml:"port"`
+	Keyspace string `yaml:"keyspace"`
+	Username string `yaml:"username"`
+	Password string `yaml:"password"`
+}
+
+func getOpts(c *cli.Context) *opts {
+
+	var opts = &opts{
+		Port:        9042,
+		Consistency: "One",
+		Keyspace:    "cherami",
+	}
+
+	if c.IsSet("zone") {
+
+		if configYaml, err := ioutil.ReadFile(c.String("config")); err == nil {
+
+			config := make(map[string]ZoneConfig)
+
+			if err = yaml.Unmarshal(configYaml, &config); err != nil {
+
+				fmt.Printf("error parsing yaml (%s): %v\n", c.String("config"), err)
+
+			} else {
+
+				if cfg, ok := config[c.String("zone")]; !ok {
+
+					fmt.Printf("config not found for zone '%v'\n", c.String("zone"))
+
+				} else {
+
+					opts.Hosts = cfg.Hosts
+					if cfg.Port != 0 {
+						opts.Port = cfg.Port
+					}
+					opts.Keyspace = cfg.Keyspace
+					opts.Username = cfg.Username
+					opts.Password = cfg.Password
+				}
+			}
+		}
+	}
+
+	if c.IsSet("hosts") {
+		opts.Hosts = c.String("hosts")
+	}
+
+	if c.IsSet("port") {
+		opts.Port = c.Int("port")
+	}
+
+	if c.IsSet("keyspace") {
+		opts.Keyspace = c.String("keyspace")
+	}
+
+	if c.IsSet("consistency") {
+		opts.Consistency = c.String("consistency")
+	}
+
+	if c.IsSet("username") {
+		opts.Username = c.String("username")
+	}
+
+	if c.IsSet("password") {
+		opts.Password = c.String("password")
+	}
+
+	if c.IsSet("timeout") {
+		opts.Timeout = c.Duration("timeout")
+	}
+
+	if c.IsSet("retries") {
+		opts.Retries = c.Int("retries")
+	}
+
+	if c.IsSet("pagesize") {
+		opts.PageSize = c.Int("pagesize")
+	}
+
+	return opts
 }
